Adds a/b buttons.
